refactor(frontend): extract YieldDistributionChart into its own module
Move the YieldDistributionChart component (with YieldContextMenuPayload and
the legend-limit constant) out of YieldOverview into its own module,
completing the page decomposition. YieldOverview is now a thin container
(~192 lines, down from ~772) that fetches data and wires up the filter bar
and chart; all chart-only imports move with the component. No behavior
change.
